analysis-scheduler/pyproject.toml
[tool.poetry]
name = "analysis_scheduler"
version = "0.1.0"
description = "Schedules analysis jobs"
authors = ["Alexandre Sparton"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.2"
pymongo = "^4.11.3"
requests = "^2.32.3"
kafka-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
